feat: add PaymentUpload component for uploading payment proofs
- Implemented PaymentUpload component with file validation, preview, and upload functionality.
- Integrated upload progress and success/error handling with user feedback via toast notifications.
- Updated order detail index to include PaymentUpload component.
- Enhanced OrdersList and OrderCard components for better order management and display.
- Added pagination and filtering capabilities to OrdersList for improved user experience.
