+++ apps/web/src/components/orders/detail/index.ts
@@ -1,0 +1,3 @@
+export { default as OrderOverview } from "./OrderOverview";
+export { default as PaymentUpload } from "./PaymentUpload";
+export { default as ConfirmButton } from "./ConfirmButton";
