+++ apps/web/src/components/orders/index.ts
@@ -1,0 +1,5 @@
+// List components
+export * from "./list";
+
+// Detail components
+export * from "./detail";
