+++ apps/web/src/components/orders/list/OrdersList.tsx
@@ -1,27 +1,20 @@
 "use client";
 
 import * as React from "react";
-import Link from "next/link";
-import Image from "next/image";
-import { useRouter } from "next/navigation";
 import { motion, AnimatePresence } from "framer-motion";
 import { format } from "date-fns";
-import { Badge } from "../../components/ui/badge";
-import { Button } from "../../components/ui/button";
-import { Calendar as CalendarComponent } from "../../components/ui/calendar";
+import { Button } from "@/components/ui/button";
+import { Calendar as CalendarComponent } from "@/components/ui/calendar";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "../../components/ui/popover";
+} from "@/components/ui/popover";
 import {
   Search,
   Calendar,
   ChevronLeft,
   ChevronRight,
-  Package,
-  Eye,
-  MoreHorizontal,
   X,
   ShoppingBag,
   Clock,
@@ -32,17 +25,13 @@ import {
   Truck,
   CreditCard,
 } from "lucide-react";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "../../components/ui/dropdown-menu";
-import DevUserSwitcher from "../../components/DevUserSwitcher";
+import DevUserSwitcher from "@/components/DevUserSwitcher";
 import { useCustomerOrders, type Order } from "@/hooks/useCustomerOrders";
+import OrderCard from "./OrderCard";
 
-export default function OrdersPage() {
-  const router = useRouter();
+export default function OrdersList() {
+  
+  
   const [searchInput, setSearchInput] = React.useState<string>("");
   const [q, setQ] = React.useState<string | null>(null);
   const [dateRange, setDateRange] = React.useState<{
@@ -73,6 +62,8 @@ export default function OrdersPage() {
   const orders: Order[] = data?.items || [];
   const total = data?.total ?? 0;
   const isPaginating = isFetching && !isLoading;
+
+  
 
   const statusConfig: Record<
     string,
@@ -198,7 +189,9 @@ export default function OrdersPage() {
           </div>
           <h2 className="text-2xl font-bold mb-2">Something went wrong</h2>
           <p className="text-muted-foreground mb-6">
-            {error instanceof Error ? error.message : "Failed to load your orders. Please try again."}
+            {error instanceof Error
+              ? error.message
+              : "Failed to load your orders. Please try again."}
           </p>
           <Button
             onClick={() => window.location.reload()}
@@ -264,7 +257,8 @@ export default function OrdersPage() {
                     <span className="truncate">
                       {dateRange.from ? (
                         dateRange.to ? (
-                          dateRange.from.getTime() === dateRange.to.getTime() ? (
+                          dateRange.from.getTime() ===
+                          dateRange.to.getTime() ? (
                             format(dateRange.from, "PPP")
                           ) : (
                             <>
@@ -291,7 +285,10 @@ export default function OrdersPage() {
                     )}
                   </Button>
                 </PopoverTrigger>
-                <PopoverContent className="w-auto p-0 rounded-2xl border-border/60 shadow-lg" align="start">
+                <PopoverContent
+                  className="w-auto p-0 rounded-2xl border-border/60 shadow-lg"
+                  align="start"
+                >
                   <div className="p-4 space-y-3">
                     <CalendarComponent
                       mode="range"
@@ -299,7 +296,9 @@ export default function OrdersPage() {
                         from: dateRange.from,
                         to: dateRange.to,
                       }}
-                      onSelect={(range: { from?: Date; to?: Date } | undefined) => {
+                      onSelect={(
+                        range: { from?: Date; to?: Date } | undefined
+                      ) => {
                         setDateRange({
                           from: range?.from,
                           to: range?.to,
@@ -428,165 +427,25 @@ export default function OrdersPage() {
 
               <AnimatePresence mode="wait">
                 <motion.div
-                  key={`${status || "all"}-${q || "noq"}-${dateRange.from?.getTime() || "nodate"}-${dateRange.to?.getTime() || "nodate"}`}
+                  key={`${status || "all"}-${q || "noq"}-${
+                    dateRange.from?.getTime() || "nodate"
+                  }-${dateRange.to?.getTime() || "nodate"}`}
                   initial={{ opacity: 0, y: 20 }}
                   animate={{ opacity: 1, y: 0 }}
                   exit={{ opacity: 0, y: -20 }}
                   transition={{ duration: 0.3 }}
                   className="space-y-3"
                 >
-                  {orders.map((order, index) => {
-                    const orderStatus = statusConfig[order.status] || {
-                      color: "bg-muted text-muted-foreground border-muted",
-                      icon: <Package className="h-3.5 w-3.5" />,
-                      label: order.status,
-                    };
-
-                    const totalItems = order.items?.reduce(
-                      (sum, it) => sum + (it?.quantity ?? 1),
-                      0
-                    ) ?? 0;
-
-                    return (
-                      <motion.div
-                        key={order.id}
-                        initial={{ opacity: 0, y: 30, scale: 0.95 }}
-                        animate={{ opacity: 1, y: 0, scale: 1 }}
-                        transition={{
-                          duration: 0.4,
-                          delay: index * 0.05,
-                          ease: [0.25, 0.46, 0.45, 0.94],
-                        }}
-                        whileHover={{
-                          y: -4,
-                          transition: { duration: 0.2 },
-                        }}
-                        className="group relative overflow-hidden rounded-xl border border-border/50 bg-gradient-to-br from-card/90 to-card/60 backdrop-blur-sm p-3 cursor-pointer"
-                        onClick={(e) => {
-                          const target = e.target as HTMLElement | null;
-                          if (!target) return;
-                          if (target.closest('button, a, input, textarea, select')) return;
-                          if (target.closest('.radix-portal')) return;
-                          router.push(`/orders/${order.id}`);
-                        }}
-                      >
-                        <div className="flex items-center justify-between mb-3 pl-3 pr-3">
-                          <h3 className="text-base lg:text-md font-semibold text-muted-foreground leading-tight">
-                            ID {order.id}
-                          </h3>
-
-                          <div>
-                            <DropdownMenu>
-                              <DropdownMenuTrigger asChild>
-                                <Button
-                                  variant="ghost"
-                                  size="sm"
-                                  className="h-8 w-8 p-0 rounded-md bg-muted/8 hover:bg-muted/16"
-                                >
-                                  <MoreHorizontal className="h-4 w-4" />
-                                </Button>
-                              </DropdownMenuTrigger>
-                              <DropdownMenuContent
-                                align="end"
-                                className="w-48"
-                              >
-                                <DropdownMenuItem asChild>
-                                  <Link
-                                    href={`/orders/${order.id}`}
-                                    className="cursor-pointer"
-                                  >
-                                    <Eye className="mr-2 h-4 w-4" />
-                                    View Details
-                                  </Link>
-                                </DropdownMenuItem>
-                                <DropdownMenuItem
-                                  onClick={() => {
-                                    navigator.clipboard.writeText(
-                                      String(order.id)
-                                    );
-                                  }}
-                                  className="cursor-pointer"
-                                >
-                                  <Package className="mr-2 h-4 w-4" />
-                                  Copy Order ID
-                                </DropdownMenuItem>
-                              </DropdownMenuContent>
-                            </DropdownMenu>
-                          </div>
-                        </div>
-
-                        <div className="rounded-lg border border-border/40 bg-[#FFFFF5]/60 backdrop-blur-sm p-3 group-hover:border-primary/20 transition-colors duration-300">
-                          <div className="flex flex-row flex-wrap items-center justify-between gap-3">
-                            <div className="flex items-center gap-3 flex-1 min-w-0">
-                              <div className="flex-shrink-0 rounded-lg overflow-hidden shadow-sm shadow-primary/20 group-hover:scale-105 transition-transform duration-300 relative h-20 w-20">
-                                {(() => {
-                                  const firstItem = order.items?.[0];
-                                  const firstProduct = firstItem?.product;
-                                  let finalImageUrl = '';
-                                  
-                                  if (Array.isArray(firstProduct?.images) && firstProduct.images.length > 0) {
-                                    const firstImageObj = firstProduct.images[0];
-                                    if (firstImageObj && typeof firstImageObj.imageUrl === 'string' && firstImageObj.imageUrl.trim()) {
-                                      finalImageUrl = firstImageObj.imageUrl.trim();
-                                    }
-                                  }
-                                  
-                                  const productName = firstProduct?.name || "Product";
-
-                                  if (finalImageUrl) {
-                                    return (
-                                      <Image
-                                        src={finalImageUrl}
-                                        alt={productName}
-                                        fill
-                                        className="object-cover"
-                                        sizes="96px"
-                                        onError={() => {}}
-                                      />
-                                    );
-                                  } else {
-                                    return (
-                                      <div className="h-full w-full p-4 bg-primary-gradient flex items-center justify-center">
-                                        <Package className="h-12 w-12 text-primary-foreground" />
-                                      </div>
-                                    );
-                                  }
-                                })()}
-                              </div>
-
-                              <div className="flex-1 min-w-0">
-                                <div className="mb-2">
-                                  <h4 className="text-base font-bold mb-1 group-hover:text-primary transition-colors line-clamp-1">
-                                    {order.items?.[0]?.product?.name || "Product Name"}
-                                  </h4>
-                                  <p className="text-sm text-muted-foreground">
-                                    {formatDate(order.createdAt)}
-                                  </p>
-                                </div>
-                              </div>
-                            </div>
-
-                            <div className="flex flex-col items-end gap-3 flex-shrink-0 min-w-[120px] sm:min-w-[140px] lg:min-w-[180px]">
-                              <Badge
-                                className={`flex items-center gap-1.5 px-3 py-1.5 border font-medium ${orderStatus.color}`}
-                              >
-                                {orderStatus.icon}
-                                {orderStatus.label}
-                              </Badge>
-                            </div>
-                          </div>
-                          <div className="flex items-center justify-between mt-4 pt-3 border-t border-border/40">
-                              <p className="text-sm font-medium text-foreground">
-                                {totalItems > 0 ? `${totalItems} ${totalItems === 1 ? 'item' : 'items'}` : '0 items'}
-                              </p>
-                            <p className="text-lg font-bold bg-gradient-to-r from-primary to-primary/70 bg-clip-text text-transparent">
-                              {formatCurrency(Number(order.grandTotal || 0))}
-                            </p>
-                          </div>
-                        </div>
-                      </motion.div>
-                    );
-                  })}
+                  {orders.map((order, index) => (
+                    <OrderCard
+                      key={order.id}
+                      order={order}
+                      index={index}
+                      statusConfig={statusConfig}
+                      formatCurrency={formatCurrency}
+                      formatDate={formatDate}
+                    />
+                  ))}
                 </motion.div>
               </AnimatePresence>
             </div>
@@ -595,7 +454,15 @@ export default function OrdersPage() {
               <div className="mt-8 p-6 rounded-2xl border border-border/50 bg-gradient-to-br from-card/90 to-card/60 backdrop-blur-sm">
                 <div className="flex flex-col sm:flex-row items-center justify-between gap-4">
                   <div className="text-sm text-muted-foreground order-2 sm:order-1">
-                    Showing <span className="font-semibold text-foreground">{orders.length}</span> of <span className="font-semibold text-foreground">{total}</span> orders
+                    Showing{" "}
+                    <span className="font-semibold text-foreground">
+                      {orders.length}
+                    </span>{" "}
+                    of{" "}
+                    <span className="font-semibold text-foreground">
+                      {total}
+                    </span>{" "}
+                    orders
                   </div>
 
                   <div className="flex items-center gap-2 order-1 sm:order-2">
