+++ apps/web/src/components/orders/list/index.ts
@@ -1,0 +1,2 @@
+export { default as OrdersList } from "./OrdersList";
+export { default as OrderCard } from "./OrderCard";
